requirements.txt
django==3.1.2
django-crispy-forms==1.9.2
django-bootstrap4==2.3.1
